Moved all binary search tree files into one folder; updated tests to use correct namespace
