+++ Algorithms/Algorithms.Tests/BinarySearchTreeTests/When_Default_Constructor_Is_Called.cs
@@ -1,3 +1,4 @@
+using Algorithms.BinarySearchTree;
 using FluentAssertions;
 using Xunit;
