+++ Algorithms/Algorithms.Tests/BinarySearchTreeTests/When_GetEnumerator_Is_Called.cs
@@ -1,3 +1,4 @@
+using Algorithms.BinarySearchTree;
 using FluentAssertions;
 using System.Collections.Generic;
 using System.Linq;
